movie/requirement.txt
Django==1.10
django-bootstrap3==9.1.0
django-filter==1.1.0
djangorestframework==3.7.7
mysqlclient==1.3.12
whitenoise==3.3.1
